Cast tls.Version* as uint16 for compatibility with tls.Config

+++ pkg/http_client.go
@@ -42,12 +42,12 @@ func (hcc *HttpClientConfig) BuildRoundTripper() (http.RoundTripper, error) {
 				return nil, fmt.Errorf("failed to add CA cert to pool: %v", hcc.AdditionalCACerts[i])
 			}
 		}
-		minVersion := tls.VersionTLS13
+		minVersion := uint16(tls.VersionTLS13)
 		switch hcc.TlsMinVersion {
 		case "1.2":
-			minVersion = tls.VersionTLS12
+			minVersion = uint16(tls.VersionTLS12)
 		case "1.3":
-			minVersion = tls.VersionTLS13
+			minVersion = uint16(tls.VersionTLS13)
 		}
 		transport.TLSClientConfig = &tls.Config{
 			ClientCAs:  certPool,
